Client-And-Server-Interaction 12/8
Setting up switch cases for client and server. From what it looks like to me, both client and server needs switch cases because they need to know based off the first 4 bytes read how many more bytes are to read and what action they need to perform. I'm not sure if an outside function can take care of this because it would have to set up the connection to the socket again, and and different actions need different variables to be passed. Lots of pseudocode and setting up my thoughts.